requirements.txt
python-telegram-bot[job-queue,webhooks]>=20.6
feedparser==6.0.11
python-dotenv==1.0.1
certifi>=2024.0.0